fixed minor sanitizer issue with unaligned memory access
when using type 1, unaligned memory is accessed through a union with ((packed)) attribute.
But the union size is still considered the size of its biggest element,
so even if we just access 2 bytes, the sanitizer expects that input can provide up to maxSize bytes.
Changed the methodology so that each type has its own size.
All warnings disappeared from tests.