renamed all *X2 decompression functions to *X1
since they decode one symbol at a time.

This used to be called *X2 because decoding tables use 2 bytes per entry.
This was to separate an earlier version which was using only 1 byte per entry,
but was slower.
This earlier version no longer exists.

The new naming scheme will make it possible and easier to introduce an *X3 variant (3 symbols per decode)